refactored project structure & tests
