+++ tests/navigation/navBar.navigation.spec.js
@@ -1,0 +1,7 @@
+// @ts-check
+const { test, expect } = require("@playwright/test");
+import { HomePage } from "../../pageObjects/home.pageObject";
+
+test.describe("Nav bar", () => {
+
+})
